feat(frames,react-frames): add frame svg formatters
