+++ packages/frames/package.json
@@ -40,9 +40,7 @@
   "module": "./build/esm/index.js",
   "main": "./build/cjs/index.js",
   "unpkg": "./build/umd/umd.min.js",
-  "peerDependencies": {
-    "@arwes/react-tools": "^1.0.0-next.4"
-  },
+  "peerDependencies": {},
   "dependencies": {
     "tslib": "^2.3.1"
   },
