+++ packages/frames/src/formatFrameSVGPolyline/index.ts
@@ -1,0 +1,1 @@
+export * from './formatFrameSVGPolyline';
