+++ packages/frames/src/types.ts
@@ -1,0 +1,5 @@
+export type FRAME_SVG_DIMENSION = number | string;
+
+export type FRAME_SVG_POINT = FRAME_SVG_DIMENSION[];
+
+export type FRAME_SVG_POLYLINE = FRAME_SVG_POINT[];
