+++ packages/react-frames/package.json
@@ -42,6 +42,7 @@
   "unpkg": "./build/umd/umd.min.js",
   "peerDependencies": {
     "@arwes/tools": "^1.0.0-next.3",
+    "@arwes/frames": "^1.0.0-next.3",
     "@arwes/react-tools": "^1.0.0-next.3",
     "react": "18"
   },
