+++ packages/react-frames/src/FrameSVG/FrameSVG.basic.sandbox.tsx
@@ -1,6 +1,6 @@
 import React, { type ReactElement } from 'react';
 import { createRoot } from 'react-dom/client';
-import { FrameSVG } from '@arwes/react-core';
+import { FrameSVG } from '@arwes/react-frames';
 
 const Sandbox = (): ReactElement => {
   return (
