+++ packages/react-frames/src/FrameSVG/index.ts
@@ -3,5 +3,5 @@ import { FrameSVG as Component } from './FrameSVG';
 
 const FrameSVG = memo(Component);
 
-export * from './FrameSVG.types';
+export * from './FrameSVG';
 export { FrameSVG };
